add: splitting code  into multilayers - updateTodo
