+++ backend/src/lambda/http/updateTodo.ts
@@ -2,20 +2,15 @@ import 'source-map-support/register'
 
 import { APIGatewayProxyEvent, APIGatewayProxyHandler, APIGatewayProxyResult } from 'aws-lambda'
 import { UpdateTodoRequest } from '../../requests/UpdateTodoRequest'
-import * as AWS from 'aws-sdk';
+import { updatedTodo } from '../../businessLogic/todos'
 
 import { createLogger } from '../../utils/logger';
-import { parseUserId } from '../../auth/utils';
 
 const logger = createLogger('updateTodos');
 
-const docClient = new AWS.DynamoDB.DocumentClient();
-
-const todosTable = process.env.TODOS_TABLE;
-
 export const handler: APIGatewayProxyHandler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   const todoId = event.pathParameters.todoId;
-  const updatedTodo: UpdateTodoRequest = JSON.parse(event.body);
+  const parsedBody: UpdateTodoRequest = JSON.parse(event.body);
 
   logger.info('Getting an item to be updated: ', {
     event
@@ -28,72 +23,14 @@ export const handler: APIGatewayProxyHandler = async (event: APIGatewayProxyEven
   const authorization = event.headers.Authorization
   const split = authorization.split(' ')
   const jwtToken = split[1]
-  const userId = parseUserId(jwtToken);
-  console.log("userId", userId)
 
-  let todoToBeUpdate = await docClient.query({
-    TableName: todosTable,
-    KeyConditionExpression: 'userId = :userId AND todoId = :todoId',
-    ExpressionAttributeValues: {
-      ':userId': userId,
-      ':todoId': todoId
-  }
-  }).promise()
-
-
-  logger.info('Item to be updated', todoToBeUpdate);
-
-  if(todoToBeUpdate.Items.length === 0) {
-    return {
-      statusCode: 404,
-      headers: {
-        'Access-Control-Allow-Origin': '*'
-      },
-      body: 'The item to be update was not found'
-    }
-  }
-
-  // I let the done property as optional
-  if(!updatedTodo.hasOwnProperty("done")) {
-    await docClient.update({
-      TableName: todosTable,
-      Key: {
-        userId,
-        todoId
-      },
-      UpdateExpression: "set #name =:name, #dueDate=:dueDate",
-      ExpressionAttributeValues: {
-        ":name": updatedTodo.name,
-        ":dueDate": updatedTodo.dueDate,
-      },
-      ExpressionAttributeNames: {"#name": "name", "#dueDate": "dueDate"},
-      ReturnValues: "UPDATED_NEW"
-    }).promise()
-
-  } else {
-    await docClient.update({
-      TableName: todosTable,
-      Key: {
-        userId,
-        todoId
-      },
-      UpdateExpression: "set #name =:name, #dueDate=:dueDate, #done=:done",
-      ExpressionAttributeValues: {
-        ":name": updatedTodo.name,
-        ":dueDate": updatedTodo.dueDate,
-        ":done": updatedTodo.done
-      },
-      ExpressionAttributeNames: {"#name": "name", "#dueDate": "dueDate", "#done": "done"},
-      ReturnValues: "UPDATED_NEW"
-    }).promise()
-  }
-
+  const result = await updatedTodo(jwtToken, todoId, parsedBody);
 
   return {
-    statusCode: 200,
+    statusCode: result.statusCode,
     headers: {
       'Access-Control-Allow-Origin': '*'
     },
-    body: ''
+    body: result.body
   }
 }
